+ Renamed `clustering` to `label_extractor` + Renamed `return_labels` to `clustering` + Moved warnings inside new `clustering` to `utils.config`
